handle if no req.files

+++ src/middlewares/validations/index.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const validation = (schema) => async (req, res, next) => {
   try {
     await schema.validate({
-      file: req.files[0],
+      file: req.files?req.files[0]:null,
       body: req.body,
       query: req.query,
       params: req.params,
@@ -11,8 +11,10 @@ const validation = (schema) => async (req, res, next) => {
     });
     return next();
   } catch (err) {
-    const pathFile = req.files[0].destination + req.files[0].filename;
-    fs.unlinkSync(pathFile)
+    if (req.files) {
+      const pathFile = req.files[0].destination + req.files[0].filename;
+      fs.unlinkSync(pathFile)
+    }
     return res
       .status(400)
       .json({ status: 400, message: err.message, data: err.name });
